[E2E] Remove `admin` spec from smoke tests (#23463)
* Remove various admin checks from smoke tests

* Move repro 22754 over to `onboarding > setup`

* Remove `admin` spec from smoke tests
